% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildGRUmaps.R
\name{Build GRU maps}
\alias{Build GRU maps}
\alias{buildGRUmaps}
\title{Build Grouped Response Units in maps}
\usage{
buildGRUmaps(gruLoc, parsValues)
}
\arguments{
\item{gruLoc}{raster file that is comprised by numbers from 1 to the number of total \code{GRUs} that were defined}

\item{parsValues}{data frame that has the values of the four parameters of each \code{GRU}. It must have equal number of
rows as number of GRU that were defined, and must have four columns which define the \code{alpha1}, \code{alpha2}, \code{d}
and \code{Smax}}
}
\value{
a list which is comprised by four raster, each one of them has the values of a parameter spatialized according with
the GRU raster layer
}
\description{
This function builds raster maps for each parameter based on a raster file where the location of the
Grouped Response Units (\code{GRUs}) are defined. This raster must have the same resolution as the forcing files
(i.e., for each cell that is planned to be simulated, there must be forcing time series and a cell assigned to a \code{GRU}).
}
\examples{
library(raster)
gru_loc <- raster("./cells_GRU/gru2_cober_location.tif")
pars <- read.csv("./param_dwb.csv")
gru_maps <- Build_gru_maps(gru_loc,pars)

}
\author{
Nicolas Duque Gardeazabal <nduqueg@unal.edu.co> 
David Zamora <dazamoraa@unal.edu.co>
Water Resources Engineering Research Group - GIREH
}
