% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sogamoso.R
\docType{data}
\name{cells}
\alias{cells}
\title{cells}
\format{The list contains
\describe{
  \item{cells}{Data frame, cells coordinates and its ID number.}
}}
\usage{
cells
}
\description{
XXX
}
\keyword{datasets}
