% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sogamoso.R
\docType{data}
\name{dwb_results}
\alias{dwb_results}
\title{dwb_results}
\format{A list with seven output variables from DWB results
\describe{
  \item{dwb_results}{List, it contains the DWB model's outputs in matrix format, q_total- total runoff,
        aet- real evapotranspiration, r- recharge, qd- Surface runoff, qd- baseflow,
        s- soil storage, g- groundwater storage.}
}}
\usage{
dwb_results
}
\description{
Results from DWB in Sogamoso River Basin
}
\keyword{datasets}
