% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sogamoso.R
\docType{data}
\name{EscSogObs}
\alias{EscSogObs}
\title{EscSogObs}
\format{
data.frame
\describe{
   \item{EscSogObs}{Data frame, it contains runoff time series measured at 32 stations within the basin.
   These gauges belong to the IDEAM monitoring network.}
}
}
\usage{
EscSogObs
}
\description{
Flow rates observed in Sogamoso River Basin at 32 gauges from January 2012 to December 2016
}
\references{
Duque-Gardeazabal, N. (2018). "Estimation of rainfall fields in data scarce colombian watersheds,
by blending remote sensed and rain gauge data, using kernel functions". Master thesis.
Universidad Nacional de Colombia, Bogotá, Colombia.
}
\keyword{datasets}
