% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDC_signatures.R
\name{FDC_signatures}
\alias{FDC_signatures}
\title{Values of Flow Duration Curve and Casper et al's Flow Duration Curve signatures.}
\usage{
FDC_signatures(data, int_func = 1)
}
\arguments{
\item{data}{A data frame with 4 columns. The first must be the observed values of the streamflow, 
the second must be the simulated values of the stream flow, the third must be the ordinal value of
the streamflow when ordered from greater to lower and the forth column must be the probability of
excedence of the streamflow.}

\item{int_func}{Optional. An integral function that has as inputs domain and range of the data to be
integrated. By defult a Gaussian cuadrature with 2 points is set.}
}
\value{
A 1 column matrix with the CDC signatures.
}
\description{
This function retrieves the Flow Duration Curve of observed and simulated values of streamflow.
Flow Duration Curves are computed with the following equation:

\deqn{F(i) = \frac{m_{i} - a}{N + 1 - 2*a}}{%
F(i) = (m_i - a)/(N + 1 - 2*a)}


Where \eqn{F(i)} is the probability of excedence of the i-th values of streamflow when ordered from greater to
lower. \eqn{m_i} is the position of the value in the ranked streamflow. \eqn{N} is the total number of streamflow values.
\eqn{a} is a parameter which depends of the probability distribution of the streamflow. 

The value of each of the 5 Flow Duration Curve signatures proposed by Casper et al (2011) are 
used to measure the similarity between two Flow Duration Curves
(tipycally one observed and one simulated).

1. BiasRR: Measure of the mean flow.
\deqn{BiasRR = (mean(FDC_sim)-mean(FDC_obs))/mean(FDC_obs) * 100 \%}

Where \eqn{mean(FDC_sim)} is the mean value of the simulated values of streamflow 
and \eqn{mean(FDC_obs)} is the mean value of the observed values of streamflow

2. BiasFDCmidslope: Measure of flashiness of the catchment.
*ecuación*

3. BiasFHV: Measure of the higer values in the FDC.
*ecuación*

4. BiasFLV: Measure of the lower values in the FDC.
*ecuación*

5. BiasMM: Measure of the median values in the FDC.
*ecuación*
}
\examples{
CDC_signatures(data)

}
\author{
Christian David Rodríguez <chdrodriguezca@unal.edu.co> \cr
David Zamora <dazamoraa@unal.edu.co> \cr

Water Resources Engineering Research Group - GIREH
Universidad Nacional de Colombia - Sede Bogota
}
