% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FU_fuction.R
\name{fun_FU}
\alias{fun_FU}
\title{ESTA ES LA FUCION DE NICO}
\usage{
fun_FU(numerador, denominador, param)
}
\arguments{
\item{numerador}{}

\item{denominador}{}

\item{param}{}
}
\description{
ESTA ES LA FUCION DE NICO
}
\examples{

}
