% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sogamoso.R
\docType{data}
\name{gru_maps}
\alias{gru_maps}
\title{gru_maps}
\format{list and raster
\describe{
   \item{alpha1}{a vector with alpha1 values for each GRU}
   \item{alpha2}{a vector with alpha2 values for each GRU}
   \item{smax}{a vector with Smax values for each GRU}
   \item{d}{a vector with d values for each GRU}
   \item{alpha1R}{a raster with alpha1 values for each GRU}
   \item{alpha2R}{a raster with alpha2 values for each GRU}
   \item{smaxR}{a raster with Smax values for each GRU}
   \item{dR}{a raster with d values for each GRU}
}}
\usage{
gru_maps
}
\description{
Spatial distribution of DWB model parameters in Sogamoso River basin
}
\keyword{datasets}
