% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sogamoso.R
\docType{data}
\name{P_sogamoso}
\alias{P_sogamoso}
\title{P_sogamoso}
\format{The list contains
\describe{
  \item{P_sogamoso}{Data frame, it should represent the cells in each row, and the precipitation
        info. by month in each column. The cell rank should match the cell ID in the cells data frame.}
}}
\usage{
P_sogamoso
}
\description{
XXX
}
\keyword{datasets}
