% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sogamoso.R
\docType{data}
\name{param}
\alias{param}
\title{param}
\format{The list contains
\describe{
  \item{param}{Data frame, it should represent a GRU in each row, and parameter values in each
        column. GRU rank should match the GRU number used in the GRU raster.}
}}
\usage{
param
}
\description{
XXX
}
\keyword{datasets}
