% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sogamoso.R
\docType{data}
\name{PET_sogamoso}
\alias{PET_sogamoso}
\title{PET_sogamoso}
\format{The list contains
\describe{
  \item{PET_sogamoso}{Data frame, it contains evapotranspiration data, following the same
         format as the P_sogamoso variable.}
}}
\usage{
PET_sogamoso
}
\description{
XXX
}
\keyword{datasets}
