% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sogamoso.R
\docType{data}
\name{setup_data}
\alias{setup_data}
\title{setup_data}
\format{The list contains
\describe{
   \item{setup_data}{Data frame, set-up and print options to run DWB model.}
}}
\usage{
setup_data
}
\description{
XXX
}
\keyword{datasets}
