% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sogamoso.R
\docType{data}
\name{simDWB.sogamoso}
\alias{simDWB.sogamoso}
\title{simDWB.sogamoso}
\format{
data.frame
\describe{
   \item{simDWB.sogamoso}{Data frame, it contains simulated runoff time series at the same 32 
   stations within the basin.}
}
}
\usage{
simDWB.sogamoso
}
\description{
Simulated runoff by the DWBmodelUN in the same stations where there were observed data from the Sogamoso basin
}
\references{
Duque-Gardeazabal, N. (2018). Estimation of rainfall fields in data scarce colombian watersheds,
by blending remote sensed and rain gauge data, using kernel functions. Master thesis.
Universidad Nacional de Colombia, Bogotá, Colombia. Retrieved from http://bdigital.unal.edu.co/71663/.
}
\keyword{datasets}
