% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sogamoso.R
\docType{data}
\name{basins}
\alias{basins}
\title{basins}
\format{The list contains
\describe{
  \item{basins}{Shapefile featuring subbasins accross the Sogamosos Basin.}
}}
\usage{
basins
}
\description{
XXXXX
}
\keyword{datasets}
