% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sogamoso.R
\docType{data}
\name{basins}
\alias{basins}
\title{basins}
\format{SpatialPolygonsDataFrame (S4)
\describe{
  \item{basins}{Shapefile featuring subbasins accross the Sogamosos Basin.}
}}
\usage{
basins
}
\description{
The polygons of the 23 subbasins accross the Sogamosos Basin
}
\references{
Duque-Gardeazabal, N. (2018). Estimation of rainfall fields in data scarce colombian watersheds,
by blending remote sensed and rain gauge data, using kernel functions. Master thesis.
Universidad Nacional de Colombia, Bogotá, Colombia.
}
\keyword{datasets}
