% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sogamoso.R
\docType{data}
\name{cells}
\alias{cells}
\title{cells}
\format{
data.frame
\describe{
  \item{cells}{Data frame (3 colums by 677 rows), cells coordinates and its ID number.}
}
}
\usage{
cells
}
\description{
Coordinates (Latitud and Longitud) and ID number of cells in Sogamoso River Basin
}
\references{
Duque-Gardeazabal, N. (2018). "Estimation of rainfall fields in data scarce colombian watersheds,
by blending remote sensed and rain gauge data, using kernel functions". Master thesis.
Universidad Nacional de Colombia, Bogotá, Colombia.
}
\keyword{datasets}
