% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphDWB.R
\name{graphDWB}
\alias{graphDWB}
\title{Graph for DWB model}
\usage{
graphDWB(q_total, date.ini, ...)
}
\arguments{
\item{q_total}{asasas}

\item{date.ini}{sasasa}

\item{...}{sasas}
}
\value{
A plot of precipitation, actual evapotranspiration and runoff
}
\description{
Function to graph dynamically the results of the DWB model
}
\examples{
# Load DWB model results
data(simDWB.sogamoso, EscSogObs)
runoff.sim <- simDWB.sogamoso[ ,1]
runoff.obs <- EscSogObs[ ,1]  
Runoff <- cbind(runoff.obs, runoff.sim) 
dates <- c(2001, 1)


graphDWB (runoff)


}
\author{
Nicolas Duque Gardeazabal <nduqueg@unal.edu.co> 
Pedro Felipe Arboleda Obando <pfarboledao@unal.edu.co> 
Carolina Vega Viviescas <cvegav@unal.edu.co>
David Zamora <dazamoraa@unal.edu.co>

Water Resources Engineering Research Group - GIREH
Universidad Nacional de Colombia - sede Bogotá
}
