% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphDWB.R
\name{graphDWB}
\alias{graphDWB}
\title{Graph for DWB model results}
\usage{
graphDWB(var, tp, main)
}
\arguments{
\item{var}{It is a list that contains time series of type "ts" which you want to graph. 
For (\emph{tp = 3}), it must first contain the observed precipitation series, 
followed by the simulated runoff series and finally the observed runoff.}

\item{tp}{Variable to choose the type of graph}

\item{main}{Main title for the graph}
}
\value{
Prints a dynamic graph according to the requirements.
}
\description{
This function dynamically graphs the inputs and results of the DWBmodelUN.
It has three types of graphs: 
The first (\emph{tp = 1}) to graph any variable in continuous format.
The second (\emph{tp = 2}) is to compare the runoff result of the model, with the observations.
Finally, (\emph{tp = 3}) allows to show a comparison between the observed and simulated runoff, as well as, with a set of precipitation.
}
\examples{
# Example 1
data(P_sogamoso)
P.est <- ts(c(t(P_sogamoso[1, -2:-1])), star = c(2001, 1), frequency = 12)
var <- list("Precipitation" = P.est)
 
graphDWB (var, tp = 1, main = "Precipitation Lat:7.0 Lon:-72.94")

# Example 2
data(simDWB.sogamoso, EscSogObs)
runoff.sim <- ts(simDWB.sogamoso[ ,1], star = c(2001, 1), frequency = 12)
runoff.obs <- ts(EscSogObs[ ,1] , star = c(2001, 1), frequency = 12)
var <- list("Runoff.sim" = runoff.sim, "Runoff.obs" = runoff.obs)
 
graphDWB (var, tp = 2, main = "Runoff: Gauge 23147020")

 

}
\author{
Nicolas Duque Gardeazabal <nduqueg@unal.edu.co> 
Pedro Felipe Arboleda Obando <pfarboledao@unal.edu.co> 
Carolina Vega Viviescas <cvegav@unal.edu.co>
David Zamora <dazamoraa@unal.edu.co>

Water Resources Engineering Research Group - GIREH
Universidad Nacional de Colombia - sede Bogotá
}
