% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sogamoso.R
\docType{data}
\name{GRU}
\alias{GRU}
\title{GRU}
\format{The list contains
\describe{
  \item{GRU}{Raster, it represents the Group Response Units across the Sogamoso Basin.}
}}
\usage{
GRU
}
\description{
XXX
}
\keyword{datasets}
