% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sogamoso.R
\docType{data}
\name{GRU}
\alias{GRU}
\title{GRU}
\format{A raster 
\describe{
  \item{GRU}{Raster, it represents the ten (10) Group Response Units across the Sogamoso River Basin.}
}}
\usage{
GRU
}
\description{
Raster data of Group Response Units in Sogamoso River Basin
}
\keyword{datasets}
