% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sogamoso.R
\docType{data}
\name{In_ground}
\alias{In_ground}
\title{In_ground}
\format{The list contains
\describe{
  \item{In_ground}{Raster, initial conditions in the soil storage.}
}}
\usage{
In_ground
}
\description{
XXX
}
\keyword{datasets}
