% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sogamoso.R
\docType{data}
\name{In_storage}
\alias{In_storage}
\title{In_storage}
\format{A raster 
\describe{
  \item{In_storage}{Raster, initial conditions in the groundwater storage.}
}}
\usage{
In_storage
}
\description{
Initial conditions in the groundwater storage in DWB model of Sogamoso River Basin
}
\keyword{datasets}
