% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_state.R
\name{init_state}
\alias{init_state}
\title{Initial conditions of the model}
\usage{
init_state(srmax, path_init)
}
\arguments{
\item{srmax}{Maximum storage in the root zone}

\item{path_init}{Directory to read the raster files \code{\emph{.tif}} file of initial storage conditions}
}
\value{
A list cointaing initial conditions in storage and in ground
}
\description{

}
\examples{

}
\references{
Budyko. (1974). Climate and life. New York: Academic Press, INC.

Zhang, L., Potter, N., Hickel, K., Zhang, Y., & Shao, Q. (2008). Water balance modeling over variable time scales based on the Budyko framework - Model development and testing. Journal of Hydrology, 360(1-4), 117-131. https://doi.org/10.1016/j.jhydrol.2008.07.021
}
\author{
Nicolas Duque Gardeazabal <nduqueg@unal.edu.co>
Pedro Felipe Arboleda <pfarboledao@unal.edu.co>
Carolina Vega Viviescas <cvegav@unal.edu.co>
David Zamora <dazamoraa@unal.edu.co>
Water Resources Engineering Research Group - GIREH
Universidad Nacional de Colombia
}
