% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_state.R
\name{init_state}
\alias{init_state}
\title{Initial conditions of the model}
\usage{
init_state(srmax, path_init)
}
\arguments{
\item{srmax}{Maximum storage in the root zone in Raster format.}

\item{path_init}{Directory to read the raster files \emph{.tif} of initial storage conditions.}
}
\value{
A list cointaing initial conditions in storage and in ground.
}
\description{
This function uploads or creates the initial conditions of the two-state variables present in the DWB model, in raster format.
}
\details{
It requires the raster composed of the Srmax values that were created using the \code{\link{buildGRUmaps}} function
and a path from where the function can read two rasters previously created. If the path or those rasters cannot
be found, the function creates those two rasters using the value of the Srmax reduced by half.
}
\references{
Budyko. (1974). Climate and life. New York: Academic Press, INC.

Zhang, L., Potter, N., Hickel, K., Zhang, Y., & Shao, Q. (2008). 
Water balance modeling over variable time scales based on the Budyko framework - Model 
development and testing. Journal of Hydrology, 360(1-4), 
117-131. https://doi.org/10.1016/j.jhydrol.2008.07.021
}
\author{
Nicolas Duque Gardeazabal <nduqueg@unal.edu.co> \cr
Pedro Felipe Arboleda <pfarboledao@unal.edu.co> \cr
Carolina Vega Viviescas <cvegav@unal.edu.co> \cr
David Zamora <dazamoraa@unal.edu.co> \cr

Water Resources Engineering Research Group - GIREH
Universidad Nacional de Colombia - Sede Bogota
}
