% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_state.R
\name{init_state}
\alias{init_state}
\title{Initial conditions of the model}
\usage{
init_state(srmax, path_init)
}
\arguments{
\item{srmax}{}

\item{path_init}{}
}
\value{
A list cointaing initial conditions in storage and in ground
}
\description{

}
\examples{

}
\author{
Nicolas Duque Gardeazabal <nduqueg@unal.edu.co>
Pedro Felipe Arboleda <pfarboledao@unal.edu.co>
David Zamora <dazamoraa@unal.edu.co>
Carolina Vega Viviescas <cvegav@unal.edu.co>
Grupo de Investigación en Ingeniería de los Recursos Hídricos - GIREH
Universidad Nacional de Colombia
}
