% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sogamoso.R
\docType{data}
\name{PET_sogamoso}
\alias{PET_sogamoso}
\title{PET_sogamoso}
\format{data.frame
\describe{
  \item{PET_sogamoso}{Data frame, it contains evapotranspiration data, representing the cells in each row, 
                 and the evapotranspiration info. by month in each column. 
                 The cell rank should match the cell ID in the cells data frame.}
}}
\usage{
PET_sogamoso
}
\description{
Distributed monthly potential evapotranspiration in Sogamoso River Basin from January 2012 to December 2016
}
\references{
Duque-Gardeazabal, N. (2018). Estimation of rainfall fields in data scarce colombian watersheds,
by blending remote sensed and rain gauge data, using kernel functions. Master thesis.
Universidad Nacional de Colombia, Bogotá, Colombia. Retrieved from http://bdigital.unal.edu.co/71663/.
}
\keyword{datasets}
