% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sogamoso.R
\docType{data}
\name{PET_sogamoso}
\alias{PET_sogamoso}
\title{PET_sogamoso}
\format{The list contains
\describe{
  \item{PET_sogamoso}{Data frame, it contains evapotranspiration data, following the same
         format as the P_sogamoso variable.}
}}
\usage{
PET_sogamoso
}
\description{
XXX
}
\references{
Duque-Gardeazabal, N. (2018). Estimation of rainfall fields in data scarce colombian watersheds,
by blending remote sensed and rain gauge data, using kernel functions. Master thesis.
Universidad Nacional de Colombia, Bogotá, Colombia. Retrieved from http://bdigital.unal.edu.co/71663/.
}
\keyword{datasets}
