% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printVar.R
\name{printVar}
\alias{printVar}
\title{Print variables of interest}
\usage{
printVar(variable, coor_cells, coord_sys, dates, as)
}
\arguments{
\item{variable}{Corresponds to the results of a specific variable of the DWBCalculator}

\item{coor_cells}{Coordinates of the cells to be extracted the results}

\item{coord_sys}{Geographic coordinate system}

\item{dates}{Dates to be extracted}

\item{as}{Option to print the results as independent 'raster' (\emph{\code{.tif}}) or in a 'NetCDF' file (\emph{\code{.nc}})}
}
\value{
It saves in a folder previously created a set of raster files with the results of the 
variable of interest
}
\description{
This function that allows to print some of the variables simulated by the DWB model
}
\examples{
library(raster)
data(dwb_results)    
printVar(dwb_results[[3]], coor_cells, "+init=epsg:4326",dates)

}
\author{
Nicolas Duque Gardeazabal <nduqueg@unal.edu.co>
Pedro Felipe Arboleda Obando <pfarboledao@unal.edu.co> 
Carolina Vega Viviescas <cvegav@unal.edu.co> 
David Zamora <dazamoraa@unal.edu.co>
Water Resources Engineering Research Group - GIREH
}
