% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printVar.R
\name{printVar}
\alias{printVar}
\title{Title}
\usage{
printVar(variable, coor_celdas, sis_cord, fechas)
}
\arguments{
\item{variable}{}

\item{coor_celdas}{}

\item{sis_cord}{}

\item{fechas}{}
}
\description{
Title
}
