% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sogamoso.R
\docType{data}
\name{r.cells}
\alias{r.cells}
\title{r.cells}
\format{The list contains
\describe{
  \item{r.cells}{Raster, data frame Cells converted to raster format.}
}}
\usage{
r.cells
}
\description{
XXX
}
\keyword{datasets}
