% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sogamoso.R
\docType{data}
\name{sogamoso}
\alias{sogamoso}
\title{Sogamoso River Basin data}
\format{An object of class \code{list} of length 11.}
\usage{
sogamoso
}
\description{
Sogamoso River Basin data
}
\keyword{datasets}
